Move component SiteTopUpper from 'components' to 'components/index'
